sql: functions treat '\0' as a usual symbol
Fixes LIKE, LENGTH, SUBSTR, GROUP_CONCAT and REPLACE
functions. '\0' is now treated as
a usual symbol. Strings with '\0' are now processed
entirely. Consider examples:

LENGTH(CHAR(65,00,65)) == 3
LIKE(CHAR(65,00,65), CHAR(65,00,66)) == False

Also the patch changes the way we count length of utf-8
strings.

Before we processed each byte of the string. Now
we use the following algorithm. Starting from the first
byte in string, we try to determine what kind of byte
it is: first byte of 1,2,3 or 4 byte sequence. Then
we skip corresponding amount of bytes and increment
symbol length (e.g. 2 bytes for 2 byte sequence). If
current byte is not a valid first byte of any sequence,
when we skip it and increment symbol length.

This new approach increases performance of LENGTH(),
INSTR() and TRIM(). We also thought about
another approach:
process all bytes one by one and raise an error any
time we encounter invalid byte sequence. We decided
to stick to the first approach because performance
in normal case is in our focus.

Closes #3542